``` refactor: Remove BatchRequests and ProxiedBatchRequests classes, simplify ProxyConfiguration
- DevBase.Net: Removed BatchRequests class with rate limiting and batch processing
- DevBase.Net: Removed ProxiedBatchRequests class with proxy rotation strategies
- DevBase.Net: Simplified ProxyConfiguration to basic proxy info holder
- DevBase.Net: Removed Batch, BatchProgressInfo, RequeueDecision nested classes
- DevBase.Net: Removed ProxyRotationStrategy enum and rotation logic
- DevBase.Net.Tests
