+++ DevBase.Net/Batch/BatchRequests.cs
@@ -1,9 +1,10 @@
 using System.Collections.Concurrent;
 using System.Net;
 using System.Text;
+using DevBase.Net.Core;
 using DevBase.Net.Utils;
 
-namespace DevBase.Net.Core;
+namespace DevBase.Net.Batch;
 
 public sealed class BatchRequests : IDisposable, IAsyncDisposable
 {
@@ -776,122 +777,4 @@ public sealed class BatchRequests : IDisposable, IAsyncDisposable
     }
 
     #endregion
-}
-
-public sealed class Batch
-{
-    private readonly ConcurrentQueue<Request> _queue = new();
-    private readonly BatchRequests _parent;
-
-    public string Name { get; }
-    public int QueueCount => _queue.Count;
-
-    internal Batch(string name, BatchRequests parent)
-    {
-        Name = name;
-        _parent = parent;
-    }
-
-    public Batch Add(Request request)
-    {
-        ArgumentNullException.ThrowIfNull(request);
-        _queue.Enqueue(request);
-        return this;
-    }
-
-    public Batch Add(IEnumerable<Request> requests)
-    {
-        foreach (Request request in requests)
-            Add(request);
-        return this;
-    }
-
-    public Batch Add(string url)
-    {
-        return Add(new Request(url));
-    }
-
-    public Batch Add(IEnumerable<string> urls)
-    {
-        foreach (string url in urls)
-            Add(url);
-        return this;
-    }
-
-    public Batch Enqueue(Request request) => Add(request);
-    public Batch Enqueue(string url) => Add(url);
-    public Batch Enqueue(IEnumerable<Request> requests) => Add(requests);
-    public Batch Enqueue(IEnumerable<string> urls) => Add(urls);
-
-    public Batch Enqueue(string url, Action<Request> configure)
-    {
-        Request request = new Request(url);
-        configure(request);
-        return Add(request);
-    }
-
-    public Batch Enqueue(Func<Request> requestFactory)
-    {
-        return Add(requestFactory());
-    }
-
-    public bool TryDequeue(out Request? request)
-    {
-        return _queue.TryDequeue(out request);
-    }
-
-    internal List<Request> DequeueAll()
-    {
-        List<Request> requests = new List<Request>(_queue.Count);
-        while (_queue.TryDequeue(out Request? request))
-            requests.Add(request);
-        return requests;
-    }
-
-    public void Clear()
-    {
-        while (_queue.TryDequeue(out _)) { }
-    }
-
-    public BatchRequests EndBatch() => _parent;
-}
-
-public sealed record BatchProgressInfo(
-    string BatchName,
-    int Completed,
-    int Total,
-    int Errors
-)
-{
-    public double PercentComplete => Total > 0 ? (double)Completed / Total * 100 : 0;
-    public int Remaining => Total - Completed;
-}
-
-public sealed record BatchStatistics(
-    int BatchCount,
-    int TotalQueuedRequests,
-    int ProcessedRequests,
-    int ErrorCount,
-    Dictionary<string, int> RequestsPerBatch
-)
-{
-    public double SuccessRate => ProcessedRequests > 0 
-        ? (double)(ProcessedRequests - ErrorCount) / ProcessedRequests * 100 
-        : 0;
-}
-
-public readonly struct RequeueDecision
-{
-    public bool ShouldRequeue { get; }
-    public Request? ModifiedRequest { get; }
-
-    private RequeueDecision(bool shouldRequeue, Request? modifiedRequest = null)
-    {
-        ShouldRequeue = shouldRequeue;
-        ModifiedRequest = modifiedRequest;
-    }
-
-    public static RequeueDecision NoRequeue => new(false);
-    public static RequeueDecision Requeue() => new(true);
-    public static RequeueDecision RequeueWith(Request modifiedRequest) => new(true, modifiedRequest);
 }
