+++ DevBase.Net/Batch/Proxied/ProxiedBatchRequests.cs
@@ -1,10 +1,12 @@
 using System.Collections.Concurrent;
 using System.Net;
 using System.Text;
+using DevBase.Net.Batch.Strategies;
+using DevBase.Net.Core;
 using DevBase.Net.Proxy;
 using DevBase.Net.Utils;
 
-namespace DevBase.Net.Core;
+namespace DevBase.Net.Batch.Proxied;
 
 public sealed class ProxiedBatchRequests : IDisposable, IAsyncDisposable
 {
@@ -964,197 +966,4 @@ public sealed class ProxiedBatchRequests : IDisposable, IAsyncDisposable
     }
 
     #endregion
-}
-
-public sealed class ProxiedBatch
-{
-    private readonly ConcurrentQueue<Request> _queue = new();
-    private readonly ProxiedBatchRequests _parent;
-
-    public string Name { get; }
-    public int QueueCount => _queue.Count;
-
-    internal ProxiedBatch(string name, ProxiedBatchRequests parent)
-    {
-        Name = name;
-        _parent = parent;
-    }
-
-    public ProxiedBatch Add(Request request)
-    {
-        ArgumentNullException.ThrowIfNull(request);
-        _queue.Enqueue(request);
-        return this;
-    }
-
-    public ProxiedBatch Add(IEnumerable<Request> requests)
-    {
-        foreach (Request request in requests)
-            Add(request);
-        return this;
-    }
-
-    public ProxiedBatch Add(string url)
-    {
-        return Add(new Request(url));
-    }
-
-    public ProxiedBatch Add(IEnumerable<string> urls)
-    {
-        foreach (string url in urls)
-            Add(url);
-        return this;
-    }
-
-    public ProxiedBatch Enqueue(Request request) => Add(request);
-    public ProxiedBatch Enqueue(string url) => Add(url);
-    public ProxiedBatch Enqueue(IEnumerable<Request> requests) => Add(requests);
-    public ProxiedBatch Enqueue(IEnumerable<string> urls) => Add(urls);
-
-    public ProxiedBatch Enqueue(string url, Action<Request> configure)
-    {
-        Request request = new Request(url);
-        configure(request);
-        return Add(request);
-    }
-
-    public ProxiedBatch Enqueue(Func<Request> requestFactory)
-    {
-        return Add(requestFactory());
-    }
-
-    public bool TryDequeue(out Request? request)
-    {
-        return _queue.TryDequeue(out request);
-    }
-
-    internal List<Request> DequeueAll()
-    {
-        List<Request> requests = new List<Request>(_queue.Count);
-        while (_queue.TryDequeue(out Request? request))
-            requests.Add(request);
-        return requests;
-    }
-
-    public void Clear()
-    {
-        while (_queue.TryDequeue(out _)) { }
-    }
-
-    public ProxiedBatchRequests EndBatch() => _parent;
-}
-
-public sealed record ProxiedBatchStatistics(
-    int BatchCount,
-    int TotalQueuedRequests,
-    int ProcessedRequests,
-    int ErrorCount,
-    int ProxyFailureCount,
-    int TotalProxies,
-    int AvailableProxies,
-    Dictionary<string, int> RequestsPerBatch
-)
-{
-    public double SuccessRate => ProcessedRequests > 0
-        ? (double)(ProcessedRequests - ErrorCount) / ProcessedRequests * 100
-        : 0;
-
-    public double ProxyAvailabilityRate => TotalProxies > 0
-        ? (double)AvailableProxies / TotalProxies * 100
-        : 0;
-}
-
-public sealed record ProxyFailureContext(
-    TrackedProxyInfo Proxy,
-    System.Exception Exception,
-    Request FailedRequest,
-    bool ProxyTimedOut,
-    int CurrentFailureCount,
-    int RemainingAvailableProxies
-);
-
-public interface IProxyRotationStrategy
-{
-    TrackedProxyInfo? SelectProxy(List<TrackedProxyInfo> proxies, ref int currentIndex);
-}
-
-public sealed class RoundRobinStrategy : IProxyRotationStrategy
-{
-    public TrackedProxyInfo? SelectProxy(List<TrackedProxyInfo> proxies, ref int currentIndex)
-    {
-        if (proxies.Count == 0)
-            return null;
-
-        int attempts = 0;
-        while (attempts < proxies.Count)
-        {
-            currentIndex = (currentIndex + 1) % proxies.Count;
-            var proxy = proxies[currentIndex];
-
-            if (proxy.IsAvailable())
-                return proxy;
-
-            attempts++;
-        }
-
-        return null;
-    }
-}
-
-public sealed class RandomStrategy : IProxyRotationStrategy
-{
-    private static readonly Random Random = new();
-
-    public TrackedProxyInfo? SelectProxy(List<TrackedProxyInfo> proxies, ref int currentIndex)
-    {
-        var available = proxies.Where(p => p.IsAvailable()).ToList();
-        if (available.Count == 0)
-            return null;
-
-        int index = Random.Next(available.Count);
-        var selected = available[index];
-        currentIndex = proxies.IndexOf(selected);
-        return selected;
-    }
-}
-
-public sealed class LeastFailuresStrategy : IProxyRotationStrategy
-{
-    public TrackedProxyInfo? SelectProxy(List<TrackedProxyInfo> proxies, ref int currentIndex)
-    {
-        var available = proxies.Where(p => p.IsAvailable()).ToList();
-        if (available.Count == 0)
-            return null;
-
-        var selected = available.OrderBy(p => p.FailureCount).ThenBy(p => p.TotalTimeouts).First();
-        currentIndex = proxies.IndexOf(selected);
-        return selected;
-    }
-}
-
-public sealed class StickyStrategy : IProxyRotationStrategy
-{
-    public TrackedProxyInfo? SelectProxy(List<TrackedProxyInfo> proxies, ref int currentIndex)
-    {
-        if (proxies.Count == 0)
-            return null;
-
-        if (currentIndex >= 0 && currentIndex < proxies.Count)
-        {
-            var current = proxies[currentIndex];
-            if (current.IsAvailable())
-                return current;
-        }
-
-        for (int i = 0; i < proxies.Count; i++)
-        {
-            if (proxies[i].IsAvailable())
-            {
-                currentIndex = i;
-                return proxies[i];
-            }
-        }
-
-        return null;
-    }
 }
