+++ DevBase.Net/Batch/Strategies/IProxyRotationStrategy.cs
@@ -1,0 +1,8 @@
+using DevBase.Net.Proxy;
+
+namespace DevBase.Net.Batch.Strategies;
+
+public interface IProxyRotationStrategy
+{
+    TrackedProxyInfo? SelectProxy(List<TrackedProxyInfo> proxies, ref int currentIndex);
+}
